refactor:Move Vue adapter to adapters and add Injector
