+++ src/index.ts
@@ -1,10 +1,15 @@
-import { VuePlugin } from './core/adapter/vue/VuePlugin';
+import { VuePlugin } from './adapters/vue/VuePlugin';
+import { Injector } from './compat/Injector';
 import { ObserverHub } from './core/hooks/ObserverHub';
-import { Injector } from './core/Injector/Injector';
 import { Action } from './core/Injector/types';
 import { createActivityStore } from './core/signal/observeActivitySignal';
 import { DOMWatcher } from './core/watcher/DomWatcher';
 
+export type {
+	VueMountArtifact,
+	VueMountHandle,
+	VueMountInstance
+} from './adapters/vue/type';
 export type {
 	AdapterMountInput,
 	AdapterMountResult,
@@ -14,11 +19,6 @@ export type {
 	MountAdapter,
 	ResolvableMountAdapter
 } from './core/adapter/types';
-export type {
-	VueMountArtifact,
-	VueMountHandle,
-	VueMountInstance
-} from './core/adapter/vue/type';
 
 export type {
 	LifecycleHookMap,
